requirements.txt
streamlit
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
requests>=2.26.0
pyyaml>=5.4.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
schedule>=1.1.0
vaderSentiment>=3.3.2
python-telegram-bot>=13.7
plotly>=5.0.0
